StudentDetailService moved to folder Student 	modified:   app/Http/Controllers/api/Student/StudentDetailController.php 	renamed:    app/Service/StudentDetailService.php -> app/Service/Student/StudentDetailService.php 	modified:   tests/Feature/Service/StudentDetailServiceTest.php 	modified:   tests/Feature/Student/StudentDetailControllerTest.php
