+++ app/Service/Student/StudentDetailService.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Service;
+namespace App\Service\Student;
 
 use App\Exceptions\StudentNotFoundException;
 use App\Exceptions\ResumeNotFoundException;
